backend/requirements.txt
# FastAPI 核心
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.23
alembic==1.12.1

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# AI 模型集成
openai==1.3.7
google-generativeai==0.3.2
requests==2.31.0

# 任务队列
celery==5.3.4
redis==5.0.1

# 数据处理
pydantic==2.5.0
pydantic-settings==2.1.0

# 工具库
python-dateutil==2.8.2
jinja2==3.1.2
aiofiles==23.2.1

# 开发工具
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# 监控和日志
structlog==23.2.0 